+++ src/components/10-atoms/heading/index.react.d.ts
@@ -1,11 +1,10 @@
 import React from 'react';
 
-export type Variant = 'primary' | 'secondary';
-export type Rank = 1 | 2 | 3 | 4 | 5 | 6;
+export type AXAHeadingSizes = 1 | 2 | 3 | 4 | 5 | 6;
 
 export interface AXAHeadingProps {
-  rank: Rank;
-  variant?: Variant;
+  secondary?: boolean;
+  size?: AXAHeadingSizes;
   children?: React.ReactNode;
 }