+++ src/components/10-atoms/text/index.react.d.ts
@@ -1,9 +1,12 @@
 import React from 'react';
 
-export type Variant = 'size-2' | 'size-3' | 'size-4' | 'bold' | 'semibold';
+export type AXATextSizes = 1 | 2 | 3 | 4;
 
 export interface AXATextProps {
-  variant?: Variant;
+  size?: AXATextSizes;
+  nowrap?: boolean;
+  bold?: boolean;
+  italic?: boolean;
   children?: React.ReactNode;
 }