+++ src/components/10-atoms/textarea/story.args.js
@@ -5,6 +5,9 @@ export const args = {
   readonly: false,
   required: false,
   invalid: false,
+  autocomplete: false,
+  counter: 'Still ##counter## characters left',
+  counterMax: 'Over character limit!',
 };
 
 export const argTypes = {
@@ -18,6 +21,7 @@ export const argTypes = {
   readonly: { control: 'boolean' },
   required: { control: 'boolean' },
   invalid: { control: 'boolean' },
+  autocomplete: { control: 'boolean' },
   counter: { control: 'text' },
   counterMax: { control: 'text' },
   maxLength: { control: 'text' },