+++ src/components/10-atoms/spinner/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.9
+
+Add missing color type in documentation.
+
 ## 3.0.0
 
 Support for IE11 has been discontinued. Therefore, we no longer transpile the code with Babel, the codebase is based on ES2019.